refactor: Validator 클래스 생성 후 Controller 내 검증 로직 리팩토링
